Appliance-Java-Basic/Appliance-OOP/src/main/java/com/vladproduction/Manufacturer.java
package com.vladproduction;

import java.util.Objects;

public class Manufacturer {
    private long id;
    private String name;

    public Manufacturer(long id, String name) {
        this.id = id;
        this.name = name;
    }

    public Manufacturer() {
    }

    /*public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }*/

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Manufacturer that = (Manufacturer) o;
        return Objects.equals(id, that.id) && Objects.equals(name, that.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name);
    }

    @Override
    public String toString() {
        return "Manufacturer{" +
                "id=" + id +
                ", name='" + name + '\'' +
                '}';
    }
}
